:truck: Move rollDie method and test to new classes
